Add interop to expose spatial to rust
