+++ src/extra/rust_interop.hpp
@@ -37,6 +37,9 @@ gfx::int4 gfx_getEguiMappedRegion(const SHVar &inputContext);
 gfx::int4 gfx_getViewport(const SHVar &graphicsContext);
 const egui::Input *gfx_getEguiWindowInputs(gfx::EguiInputTranslator *translator, const SHVar &graphicsContext,
                                            const SHVar &inputContext, float scalingFactor);
+
+SHTypeInfo *spatial_getSpatialContextType();
+const char *spatial_getSpatialContextVarName();
 }
 
 #endif /* E325D8E3_F64E_413D_965B_DF275CF18AC4 */
